Trying (&failing) to make it faster; streamlining output.
Have been struggling to get the updating of the current multiple-map to
work so I don't have to keep recalculating the event counts (which is
slow), but haven't figure out what I'm doing wrong.  Meantime, sticking
with the slow way.

Have changed the output to report only the inferred events (unless
_verbose flag is set on CLI), so these can be appended to a results
file.