Changed calculation of new duplication cost on move.
Have elected to change the calculation of new duplication cost to be:
(change in dup height of destination) - (change in dup height of
origin).
This is done by re-calculating the dup heights with and without the
moved p node so it's not perfectly efficient, but it appears to be
working.

Have also updated the progress message to be a bit more informative.